Split error_helpers file into 3
