+++ src/core/friendly_errors/validate_params.js
@@ -1,251 +1,16 @@
 /**
  * @for p5
  * @requires core
- *
- * This is the main file for the Friendly Error System (FES). Here is a
- * brief outline of the functions called in this system.
- *
- * The FES may be invoked by a call to either (1) _validateParameters,
- * (2) _friendlyFileLoadError, (3) _friendlyError, or (4) helpForMisusedAtTopLevelCode.
- *
- * helpForMisusedAtTopLevelCode is called by this file on window load to check for use
- * of p5.js functions outside of setup() or draw()
- * Items 1-3 above are called by functions in the p5 library located in other files.
- *
- * _friendlyFileLoadError is called by the loadX() methods.
- * _friendlyError can be called by any function to offer a helpful error message.
- *
- * _validateParameters is called by functions in the p5.js API to help users ensure
- * ther are calling p5 function with the right parameter types. The property
- * disableFriendlyErrors = false can be set from a p5.js sketch to turn off parameter
- * checking. The call sequence from _validateParameters looks something like this:
- *
- * _validateParameters
- *   lookupParamDoc
- *   scoreOverload
- *     testParamTypes
- *     testParamType
- *   getOverloadErrors
- *   _friendlyParamError
- *     ValidationError
- *     report
- *       friendlyWelcome
- *
- * The call sequences to _friendlyFileLoadError and _friendlyError are like this:
- * _friendlyFileLoadError
- *   report
- *
- * _friendlyError
- *   report
- *
- * report() is the main function that prints directly to console with the output
- * of the error helper message. Note: friendlyWelcome() also prints to console directly.
  */
-import p5 from './main';
-import * as constants from './constants';
-import { translator } from './internationalization';
-
-// p5.js blue, p5.js orange, auto dark green; fallback p5.js darkened magenta
-// See testColors below for all the color codes and names
-const typeColors = ['#2D7BB6', '#EE9900', '#4DB200', '#C83C00'];
+import p5 from '../main';
+import * as constants from '../constants';
 
 if (typeof IS_MINIFIED !== 'undefined') {
-  p5._validateParameters = p5._friendlyFileLoadError = p5._friendlyError = p5._clearValidateParamsCache = () => {};
+  p5._validateParameters = p5._clearValidateParamsCache = () => {};
 } else {
-  let doFriendlyWelcome = false; // TEMP until we get it all working LM
   // for parameter validation
-  const dataDoc = require('../../docs/parameterData.json');
+  const dataDoc = require('../../../docs/parameterData.json');
   const arrDoc = JSON.parse(JSON.stringify(dataDoc));
-
-  // -- Borrowed from jQuery 1.11.3 --
-  const class2type = {};
-  const toString = class2type.toString;
-  const names = [
-    'Boolean',
-    'Number',
-    'String',
-    'Function',
-    'Array',
-    'Date',
-    'RegExp',
-    'Object',
-    'Error'
-  ];
-  for (let n = 0; n < names.length; n++) {
-    class2type[`[object ${names[n]}]`] = names[n].toLowerCase();
-  }
-  const getType = obj => {
-    if (obj == null) {
-      return `${obj}`;
-    }
-    return typeof obj === 'object' || typeof obj === 'function'
-      ? class2type[toString.call(obj)] || 'object'
-      : typeof obj;
-  };
-
-  // -- End borrow --
-
-  const friendlyWelcome = () => {
-    // p5.js brand - magenta: #ED225D
-    //const astrixBgColor = 'transparent';
-    //const astrixTxtColor = '#ED225D';
-    //const welcomeBgColor = '#ED225D';
-    //const welcomeTextColor = 'white';
-    const welcomeMessage = translator('fes.pre', {
-      message: translator('fes.welcome')
-    });
-    console.log(
-      '    _ \n' +
-        ' /\\| |/\\ \n' +
-        " \\ ` ' /  \n" +
-        ' / , . \\  \n' +
-        ' \\/|_|\\/ ' +
-        '\n\n' +
-        welcomeMessage
-    );
-  };
-
-  /**
-   * Prints out a fancy, colorful message to the console log
-   *
-   * @method report
-   * @private
-   * @param  {String}               message the words to be said
-   * @param  {String}               func    the name of the function to link
-   * @param  {Number|String} color   CSS color string or error type
-   *
-   * @return console logs
-   */
-  const report = (message, func, color) => {
-    if (doFriendlyWelcome) {
-      friendlyWelcome();
-      doFriendlyWelcome = false;
-    }
-    if ('undefined' === getType(color)) {
-      color = '#B40033'; // dark magenta
-    } else if (getType(color) === 'number') {
-      // Type to color
-      color = typeColors[color];
-    }
-    if (func.substring(0, 4) === 'load') {
-      console.log(translator('fes.pre', { message }));
-    } else {
-      console.log(
-        translator('fes.pre', {
-          message: `${message} (http://p5js.org/reference/#p5/${func})`
-        })
-      );
-    }
-  };
-
-  // mapping used by `_friendlyFileLoadError`
-  const fileLoadErrorCases = (num, filePath) => {
-    const suggestion = translator('fes.fileLoadError.suggestion', {
-      filePath,
-      link: 'https://github.com/processing/p5.js/wiki/Local-server'
-    });
-    switch (num) {
-      case 0:
-        return {
-          message: translator('fes.fileLoadError.image', {
-            suggestion
-          }),
-          method: 'loadImage'
-        };
-      case 1:
-        return {
-          message: translator('fes.fileLoadError.xml', {
-            suggestion
-          }),
-          method: 'loadXML'
-        };
-      case 2:
-        return {
-          message: translator('fes.fileLoadError.table', {
-            suggestion
-          }),
-          method: 'loadTable'
-        };
-      case 3:
-        return {
-          message: translator('fes.fileLoadError.strings', {
-            suggestion
-          }),
-          method: 'loadStrings'
-        };
-      case 4:
-        return {
-          message: translator('fes.fileLoadError.font', {
-            suggestion
-          }),
-          method: 'loadFont'
-        };
-      case 5:
-        return {
-          message: translator('fes.fileLoadError.json', {
-            suggestion
-          }),
-          method: 'loadJSON'
-        };
-      case 6:
-        return {
-          message: translator('fes.fileLoadError.bytes', {
-            suggestion
-          }),
-          method: 'loadBytes'
-        };
-      case 7:
-        return {
-          message: translator('fes.fileLoadError.large'),
-          method: 'loadX'
-        };
-      case 8:
-        return {
-          message: translator('fes.fileLoadError.gif'),
-          method: 'loadImage'
-        };
-    }
-  };
-
-  /**
-   * This is called internally if there is a error during file loading.
-   *
-   * @method _friendlyFileLoadError
-   * @private
-   * @param  {Number} errorType
-   * @param  {String} filePath
-   */
-  p5._friendlyFileLoadError = function(errorType, filePath) {
-    const { message, method } = fileLoadErrorCases(errorType, filePath);
-    report(message, method, 3);
-  };
-
-  /**
-   * This is a generic method that can be called from anywhere in the p5
-   * library to alert users to a common error.
-   *
-   * @method _friendlyError
-   * @private
-   * @param  {Number} message message to be printed
-   * @param  {String} method name of method
-   */
-  p5._friendlyError = function(message, method) {
-    report(message, method);
-  };
-
-  /**
-   * This is called internally if there is a error with autoplay.
-   *
-   * @method _friendlyAutoplayError
-   * @private
-   */
-  p5._friendlyAutoplayError = function(src) {
-    const message = translator('fes.autoplay', {
-      src,
-      link: 'https://developer.mozilla.org/docs/Web/Media/Autoplay_guide'
-    });
-    console.log(translator('fes.pre', { message }));
-  };
 
   const docCache = {};
   const builtinTypes = new Set([
@@ -771,7 +536,7 @@ if (typeof IS_MINIFIED !== 'undefined') {
         }
       } catch (err) {}
 
-      report(`${message}.`, func, 3);
+      p5._friendlyError(`${message}.`, func, 3);
     }
   };
 
@@ -858,157 +623,7 @@ if (typeof IS_MINIFIED !== 'undefined') {
       }
     }
   };
-
-  /**
-   * Prints out all the colors in the color pallete with white text.
-   * For color blindness testing.
-   */
-  /* function testColors() {
-    const str = 'A box of biscuits, a box of mixed biscuits and a biscuit mixer';
-    report(str, 'print', '#ED225D'); // p5.js magenta
-    report(str, 'print', '#2D7BB6'); // p5.js blue
-    report(str, 'print', '#EE9900'); // p5.js orange
-    report(str, 'print', '#A67F59'); // p5.js light brown
-    report(str, 'print', '#704F21'); // p5.js gold
-    report(str, 'print', '#1CC581'); // auto cyan
-    report(str, 'print', '#FF6625'); // auto orange
-    report(str, 'print', '#79EB22'); // auto green
-    report(str, 'print', '#B40033'); // p5.js darkened magenta
-    report(str, 'print', '#084B7F'); // p5.js darkened blue
-    report(str, 'print', '#945F00'); // p5.js darkened orange
-    report(str, 'print', '#6B441D'); // p5.js darkened brown
-    report(str, 'print', '#2E1B00'); // p5.js darkened gold
-    report(str, 'print', '#008851'); // auto dark cyan
-    report(str, 'print', '#C83C00'); // auto dark orange
-    report(str, 'print', '#4DB200'); // auto dark green
-  } */
-
   p5.prototype._validateParameters = p5.validateParameters;
-}
-
-// This is a lazily-defined list of p5 symbols that may be
-// misused by beginners at top-level code, outside of setup/draw. We'd like
-// to detect these errors and help the user by suggesting they move them
-// into setup/draw.
-//
-// For more details, see https://github.com/processing/p5.js/issues/1121.
-let misusedAtTopLevelCode = null;
-const FAQ_URL =
-  'https://github.com/processing/p5.js/wiki/p5.js-overview#why-cant-i-assign-variables-using-p5-functions-and-variables-before-setup';
-
-const defineMisusedAtTopLevelCode = () => {
-  const uniqueNamesFound = {};
-
-  const getSymbols = obj =>
-    Object.getOwnPropertyNames(obj)
-      .filter(name => {
-        if (name[0] === '_') {
-          return false;
-        }
-        if (name in uniqueNamesFound) {
-          return false;
-        }
-
-        uniqueNamesFound[name] = true;
-
-        return true;
-      })
-      .map(name => {
-        let type;
-
-        if (typeof obj[name] === 'function') {
-          type = 'function';
-        } else if (name === name.toUpperCase()) {
-          type = 'constant';
-        } else {
-          type = 'variable';
-        }
-
-        return { name, type };
-      });
-
-  misusedAtTopLevelCode = [].concat(
-    getSymbols(p5.prototype),
-    // At present, p5 only adds its constants to p5.prototype during
-    // construction, which may not have happened at the time a
-    // ReferenceError is thrown, so we'll manually add them to our list.
-    getSymbols(require('./constants'))
-  );
-
-  // This will ultimately ensure that we report the most specific error
-  // possible to the user, e.g. advising them about HALF_PI instead of PI
-  // when their code misuses the former.
-  misusedAtTopLevelCode.sort((a, b) => b.name.length - a.name.length);
-};
-
-const helpForMisusedAtTopLevelCode = (e, log) => {
-  if (!log) {
-    log = console.log.bind(console);
-  }
-
-  if (!misusedAtTopLevelCode) {
-    defineMisusedAtTopLevelCode();
-  }
-
-  // If we find that we're logging lots of false positives, we can
-  // uncomment the following code to avoid displaying anything if the
-  // user's code isn't likely to be using p5's global mode. (Note that
-  // setup/draw are more likely to be defined due to JS function hoisting.)
-  //
-  //if (!('setup' in window || 'draw' in window)) {
-  //  return;
-  //}
-
-  misusedAtTopLevelCode.some(symbol => {
-    // Note that while just checking for the occurrence of the
-    // symbol name in the error message could result in false positives,
-    // a more rigorous test is difficult because different browsers
-    // log different messages, and the format of those messages may
-    // change over time.
-    //
-    // For example, if the user uses 'PI' in their code, it may result
-    // in any one of the following messages:
-    //
-    //   * 'PI' is undefined                           (Microsoft Edge)
-    //   * ReferenceError: PI is undefined             (Firefox)
-    //   * Uncaught ReferenceError: PI is not defined  (Chrome)
-
-    if (e.message && e.message.match(`\\W?${symbol.name}\\W`) !== null) {
-      const symbolName =
-        symbol.type === 'function' ? `${symbol.name}()` : symbol.name;
-      if (typeof IS_MINIFIED !== 'undefined') {
-        log(
-          `Did you just try to use p5.js's ${symbolName} ${
-            symbol.type
-          }? If so, you may want to move it into your sketch's setup() function.\n\nFor more details, see: ${FAQ_URL}`
-        );
-      } else {
-        log(
-          translator('fes.misusedTopLevel', {
-            symbolName,
-            symbolType: symbol.type,
-            link: FAQ_URL
-          })
-        );
-      }
-      return true;
-    }
-  });
-};
-
-// Exposing this primarily for unit testing.
-p5.prototype._helpForMisusedAtTopLevelCode = helpForMisusedAtTopLevelCode;
-
-if (document.readyState !== 'complete') {
-  window.addEventListener('error', helpForMisusedAtTopLevelCode, false);
-
-  // Our job is only to catch ReferenceErrors that are thrown when
-  // global (non-instance mode) p5 APIs are used at the top-level
-  // scope of a file, so we'll unbind our error listener now to make
-  // sure we don't log false positives later.
-  window.addEventListener('load', () => {
-    window.removeEventListener('error', helpForMisusedAtTopLevelCode, false);
-  });
 }
 
 export default p5;
